requirements.txt
pypng==0.20220715.0
qrcode==7.4.2
speedtest-cli==2.1.3
typing_extensions==4.8.0
